Refactor SimpleDropdownMenu items to a separate component
So we can reuse these with different menu buttons.
